Add rewad everyday bonus

+++ LukyNow/Assets/LuckyNumbers/Scripts/RewardInLuckyNumbers.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class RewardInLuckyNumbers : Reward
 {
-    public RewardInLuckyNumbers(double coin, int dollar)
+    public RewardInLuckyNumbers(double coin, double dollar)
     {
         _rewardCoin = coin;
         _rewardDollars = dollar;
